+++ Source/WebCore/html/CanvasBase.h
@@ -36,6 +36,7 @@ class AffineTransform;
 class CanvasBase;
 class CanvasRenderingContext;
 class Element;
+class GraphicsClient;
 class GraphicsContext;
 class GraphicsContextStateSaver;
 class Image;
@@ -113,12 +114,21 @@ public:
     virtual GraphicsContext* drawingContext() const;
     virtual GraphicsContext* existingDrawingContext() const;
 
+    GraphicsClient* graphicsClient() const;
+
     virtual void didDraw(const std::optional<FloatRect>&) = 0;
 
     virtual Image* copiedImage() const = 0;
     virtual void clearCopiedImage() const = 0;
 
     bool hasActiveInspectorCanvasCallTracer() const;
+
+    bool shouldAccelerate(const IntSize&) const;
+    bool shouldAccelerate(unsigned area) const;
+
+    WEBCORE_EXPORT static size_t maxActivePixelMemory();
+    WEBCORE_EXPORT static void setMaxPixelMemoryForTesting(std::optional<size_t>);
+    WEBCORE_EXPORT static void setMaxCanvasAreaForTesting(std::optional<size_t>);
 
 protected:
     explicit CanvasBase(IntSize);
@@ -132,6 +142,8 @@ protected:
     static size_t activePixelMemory();
 
     void resetGraphicsContextState() const;
+
+    void createImageBuffer(bool usesDisplayListDrawing, bool avoidBackendSizeCheckForTesting) const;
 
 private:
     virtual void createImageBuffer() const { }
